+++ src/panels/config/automation/option/ha-automation-option-row.ts
@@ -31,7 +31,7 @@ import type {
   OptionSidebarConfig,
 } from "../../../../data/automation";
 import { describeCondition } from "../../../../data/automation_i18n";
-import { fullEntitiesContext } from "../../../../data/context/context";
+import { fullEntitiesContext } from "../../../../data/context";
 import type { EntityRegistryEntry } from "../../../../data/entity/entity_registry";
 import type { Action, Option } from "../../../../data/script";
 import { showPromptDialog } from "../../../../dialogs/generic/show-dialog-box";
