+++ src/panels/config/integrations/integration-panels/zwave_js/dialog-zwave_js-rebuild-network-routes/dialog-zwave_js-rebuild-network-routes-detail.ts
@@ -22,7 +22,7 @@ import {
   internationalizationContext,
   registriesContext,
   statesContext,
-} from "../../../../../../data/context/context";
+} from "../../../../../../data/context";
 import {
   getDeviceEntityDisplayLookup,
   type DeviceRegistryEntry,
