+++ src/panels/config/integrations/integration-panels/zwave_js/dialog-zwave_js-rebuild-network-routes/dialog-zwave_js-rebuild-network-routes.ts
@@ -22,7 +22,7 @@ import "../../../../../../components/progress/ha-progress-bar";
 import {
   connectionContext,
   internationalizationContext,
-} from "../../../../../../data/context/context";
+} from "../../../../../../data/context";
 import type {
   ZWaveJSNetwork,
   ZWaveJSRebuildRoutesStatusMessage,
