+++ src/panels/config/scene/ha-scene-dashboard.ts
@@ -62,10 +62,7 @@ import {
   subscribeCategoryRegistry,
 } from "../../../data/category_registry";
 import type { CloudStatus } from "../../../data/cloud";
-import {
-  fullEntitiesContext,
-  labelsContext,
-} from "../../../data/context/context";
+import { fullEntitiesContext, labelsContext } from "../../../data/context";
 import type { DataTableFilters } from "../../../data/data_table_filters";
 import {
   deserializeFilters,
