+++ src/panels/lovelace/cards/hui-button-card.ts
@@ -26,7 +26,7 @@ import {
   entitiesContext,
   statesContext,
   uiContext,
-} from "../../../data/context/context";
+} from "../../../data/context";
 import type { EntityRegistryDisplayEntry } from "../../../data/entity/entity_registry";
 import type { ActionHandlerEvent } from "../../../data/lovelace/action_handler";
 import type { Themes } from "../../../data/ws-themes";
